backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
minio
celery
redis
python-jose[cryptography]
passlib[bcrypt]
email-validator
